% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airportGameForSampling.R
\name{airportGameForSampling}
\alias{airportGameForSampling}
\title{Construct Airport Game}
\usage{
airportGameForSampling(costs)
}
\arguments{
\item{costs}{The vector where \code{costs[i]} gives the cost for player \code{i}. \code{N} is \verb{1..length(costs)}.
\code{cost[i]} has to be smaller than or equal to \code{cost[i+1]}.}
}
\value{
The characteristic function configured with the provided cost function that returns the cost the provided set of players (S) has to pay.
}
\description{
Returns the characteristic function of an airport game with the specified cost vector.
}
\examples{
v <- airportGameForSampling(c(1, 1, 1, 3, 3, 5, 5, 5, 6, 6, 10, 10))
v(c(1, 3, 6)) # returns 5
}
\author{
David Ulrich

Alexander Rießbeck

Robin Liebermann

Mithunan Mehanathan
}
