% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approOwen.R
\name{approOwen}
\alias{approOwen}
\title{Appro Owen}
\usage{
approOwen(n, m, v, P)
}
\arguments{
\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{m}{The number of samples}

\item{v}{The characteristic function.}

\item{P}{A priori unions}
}
\value{
A list of Owen values of every player
}
\description{
Approximates the Owen value by using a simple Monte Carlo simulation.
}
\details{
It is an adaptation of an analogous procedure
for the estimation of the Shapley value and it is particularly useful when dealing with games having large sets of players.
Based on: "Estimation of the Owen Value Based on Sampling" (Alejandro Saavedra-Nieve et al., 2018)
}
\examples{
print(approOwen(3, 100, gloveGameForSampling(1:2, 3:3), list(c(1, 2), (3))))
}
\references{
Alejandro Saavedra-Nieves, M. G.-J., Ignacio García-Jurado. (2018). Estimation of the Owen Value Basedon Sampling. In E. Gil, E. Gil, J. Gil, & M. Á. Gil (Eds.), The Mathematics of the Uncertain. doi:10.1007/978-3-319-73848-2
}
\author{
Matthias Schöner
}
