% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approOwenAntithetic.R
\name{approOwenAntithetic}
\alias{approOwenAntithetic}
\title{Owen value approximation by antithetic sampling}
\usage{
approOwenAntithetic(n, m, v, P)
}
\arguments{
\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}

\item{P}{list: A priori unions.}
}
\value{
A list of Owen values of every player.
}
\description{
\code{approOwenAntithetic} is a sampling algorithm to estimate the
Owen value for all players for a specified TU game with a system of
a priori unions by using antithetic sampling.
}
\details{
This algorithm approximates the Owen value for all players by random antithetic
sampling. It is based on "Estimation of the Owen Value Based on Sampling" by Alejandro
Saavedra-Nieves et al. (2018) and was extended to make use of antithetic sampling by Staudacher and
Pollmann (2023). The base algorithm can also be found in this package
under the name \code{approOwen}.
}
\examples{
approOwenAntithetic(3, 1000, gloveGameForSampling(1:2, 3:3), list(c(1, 2), c(3)))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Approximating cooperative games using antithetic sampling",  AppliedMath
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
