% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approShapley.R
\name{approShapley}
\alias{approShapley}
\title{Appro Shapley}
\usage{
approShapley(n, m, v)
}
\arguments{
\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{m}{The number of samples}

\item{v}{The characteristic function.}
}
\value{
A list of Shapley values of every player
}
\description{
Approximates the Shapley value by using a simple Monte Carlo simulation.
}
\details{
Appro Shapley is a sampling method, which samples permutations of the players.
Based on: "Polynomial calculation of the Shapley value based on sampling" (Javier Castro Et al., 2008)
}
\examples{
print(approShapley(10, 10000, gloveGameForSampling(1:5, 6:10)))
}
\references{
CASTRO, J. et al. (2008) "Polynomial calculation of the Shapley value based on sampling", Computers and Operations Research 36, pp. 1727
}
\author{
Julian Musch

David Ulrich

Tim Pollmann

Elias Wucher

Matthias Schöner
}
