% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approShapleyAntithetic.R
\name{approShapleyAntithetic}
\alias{approShapleyAntithetic}
\title{A antithetic sampling procedure to estimate the Shapley value}
\usage{
approShapleyAntithetic(n, m, v)
}
\arguments{
\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}
}
\value{
A list of Shapley values of every player.
}
\description{
approShapleyAntithetic is a sampling methodology to estimate the
Shapley value for all players for a given TU game by antithetic sampling.
}
\details{
This algorithm approximates the Shapley value for all players by random
antithetic sampling. It is based on "Polynomial calculation of the Shapley value based on sampling"
by Javier Castro Et al. (2008) and was extended to make use of antithetic
sampling by Staudacher and Pollmann (2023). The base algorithm can also be found in this package
under the name \code{approShapley}.
}
\examples{
approShapleyAntithetic(3, 200, gloveGameForSampling(1:2, 3:3))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Approximating cooperative games using antithetic sampling",  AppliedMath
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
