% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_conf}
\alias{check_conf}
\title{Check parameter conf}
\usage{
check_conf(conf)
}
\arguments{
\item{conf}{Confidence of the true (exact) value being in the confidence interval}
}
\value{
Returns NULL
}
\description{
Check parameter conf for validity.
}
\author{
Tim Pollmann
}
