% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_i}
\alias{check_i}
\title{Check parameter i}
\usage{
check_i(i)
}
\arguments{
\item{i}{The player i}
}
\value{
Returns
}
\description{
Check parameter i for validity. Parameter i must not be NULL, must be a number
and it has to be greater than zero
}
\author{
Elias Wucher
}
