% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_m}
\alias{check_m}
\title{Check parameter m}
\usage{
check_m(m)
}
\arguments{
\item{m}{The number of samples}
}
\value{
Returns
}
\description{
Check parameter m for validity. The parameter must not be null, must be a
number and has to be greater than 0
}
\author{
Elias Wucher
}
