% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_m_n}
\alias{check_m_n}
\title{Check parameters \code{m} and \code{n}}
\usage{
check_m_n(m, n, m_max = NULL)
}
\arguments{
\item{m}{integer: The number of samples.}

\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{m_max}{Maximal value m can take}
}
\value{
None
}
\description{
Check parameters \code{m} and \code{n} for validity.
}
\author{
Tim Pollmann
}
