% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_n}
\alias{check_n}
\title{Check parameter n}
\usage{
check_n(n)
}
\arguments{
\item{n}{Indicates the cardinality of the set of players N (The number of players).}
}
\value{
Returns
}
\description{
Check parameter n for validity. Parameter n must not be NULL, must be a number
and it has to be greater than zero
}
\author{
Elias Wucher
}
