% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_P}
\alias{check_P}
\title{Check parameter P}
\usage{
check_P(P)
}
\arguments{
\item{P}{list: a priori unions}
}
\value{
Returns
}
\description{
Check parameter P for validity. Parameter P must not be NULL, must be a
list of vectors or a list of lists for the priori unions. The priori unions
must be numeric and positive. Besides the priori unions has to be disjunct.
}
\author{
Elias Wucher
}
