% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_P_i}
\alias{check_P_i}
\title{Check parameter P and i}
\usage{
check_P_i(P, i)
}
\arguments{
\item{P}{list: a priori unions}

\item{i}{integer: the player i}
}
\value{
Returns
}
\description{
Check parameter P and i for validity with the intended functions for that purpose.
Additionally it checks if player i is element of a priori union in P
}
\author{
Elias Wucher
}
