% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_positive_integer}
\alias{check_positive_integer}
\title{Check if the parameter is an integer and positive.}
\usage{
check_positive_integer(parameter)
}
\arguments{
\item{parameter}{Parameter to check}
}
\value{
None
}
\description{
Checks if the value of the parameter is an integer and positive. Note: The
datatype will not be checked. It will only be checked if the value of the
parameter is an integer.
}
\author{
Tim Pollmann
}
