% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_positive_number}
\alias{check_positive_number}
\title{Check if the parameter is a number and positive.}
\usage{
check_positive_number(parameter)
}
\arguments{
\item{parameter}{value to check}
}
\value{
None
}
\description{
Check if the parameter is a number and positive. It can be used for the
parameters \code{n}, \code{i}, \code{m} and \code{w}.
}
\author{
Elias Wucher
}
