% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_w}
\alias{check_w}
\title{Check parameter w}
\usage{
check_w(w)
}
\arguments{
\item{w}{Width of the confidence interval}
}
\value{
Returns
}
\description{
Check parameter m for validity. The parameter must not be null, must be a
number and has to be between 0 and 1
}
\author{
Elias Wucher
}
