% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitionFromIndex.R
\name{coalitionFromIndex}
\alias{coalitionFromIndex}
\title{Get a coalition from a index}
\usage{
coalitionFromIndex(n, index)
}
\arguments{
\item{n}{List of players}

\item{index}{Index describing the coalition}
}
\value{
A coalition from n described by index
}
\description{
Gets the coalition of players in list n described by the index.
}
