% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBanzhaf.R
\name{confidenceBanzhaf}
\alias{confidenceBanzhaf}
\title{Confidence Banzhaf}
\usage{
confidenceBanzhaf(i, n, v, conf, w)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{v}{function: The characteristic function.}

\item{conf}{numeric: Confidence (probability) of the true (exact) value being in the provided interval.}

\item{w}{numeric: Width of the confidence interval.}
}
\value{
Confidence interval as a vector of size 2.
}
\description{
Monte Carlo Simulation for approximating the Banzhaf value in a simple game.
}
\details{
This function calculates an approximation of the Banzhaf value. You do not
have to provide a sample size. The needed sample size is calculated inside
the function based on the given confidence and interval width. The interval
width is the width of the interval of the approximated Banzhaf value
(accuracy) while the confidence is the probability of the true (exact)
Banzhaf value being in this interval. Decreasing the interval width and/ or
increasing the confidence will lead to an increase of the required samples
and a longer computation time.
The method is based on the paper "Approximating power indices: theoretical
and empirical analysis" by Bachrach et al. from 2010.
}
\examples{
confidenceBanzhaf(3, 4, weightedVotingGameForSampling(c(1, 1, 2, 3), 1 / 2), 0.95, 0.01)
\donttest{
weights <- c(1, 1, 2, 3, 2, 3, 4, 2, 1, 1, 1, 2, 2, 3, 2, 2, 1, 2, 3)
v <- weightedVotingGameForSampling(weights, 1 / 2)
Bz <- rep(0, length(weights))
for (i in 1:length(weights)) {
  Bz[i] <- mean(confidenceBanzhaf(i, length(weights), v, 0.95, 0.03))
}
}
}
\references{
Bachrach, Y. et al. (2010) "Approximating power indices: theoretical and empirical analysis", Auton Agent Multi-Agent Syst 20, pp. 105-122
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
