% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceShapleyShubik.R
\name{confidenceShapleyShubik}
\alias{confidenceShapleyShubik}
\title{Monte Carlo Simulation for calculating the Shapley-Shubik power index in a simple game based on the paper "Approximating power indices: theoretical and empirical analysis" by Bachrach et al. from 2010}
\usage{
confidenceShapleyShubik(i, n, v, conf, w)
}
\arguments{
\item{i}{Player}

\item{n}{Number of players}

\item{v}{Characteristic function}

\item{conf}{Confidence of the true Shapley-Shubik power index being in the confidence interval}

\item{w}{Width of the confidence interval}
}
\value{
Confidence interval as a vector of size 2
}
\description{
Monte Carlo Simulation for calculating the Shapley-Shubik power index in a simple game based on the paper "Approximating power indices: theoretical and empirical analysis" by Bachrach et al. from 2010
}
\examples{
confidenceShapleyShubik(3, 4, weightedVotingGame(c(1, 1, 2, 3), 1 / 2), 0.95, 0.01)
}
