% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{createMessage}
\alias{createMessage}
\title{Creates a named list to store the info, warning and error messages used by checkCommonParams}
\usage{
createMessage(tc, msg)
}
\arguments{
\item{tc}{typeCode}

\item{msg}{message}
}
\value{
list with 2 elements. First one is called typeCode which contains an integer
representing the type of the message. Second one is the message
}
\description{
Returns a named list which includes the typeCode an the message
There are three message types:
 - info: 1
 - warning: 2
 - error: 3
}
\examples{
#Fehlt noch
}
\author{
Elias Wucher
}
