% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergodicSamplingGreedyK2.R
\name{ergodicSamplingGreedyK2}
\alias{ergodicSamplingGreedyK2}
\title{Ergodic Sampling with GreedyK2}
\usage{
ergodicSamplingGreedyK2(i, n, v, m, m1)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{v}{function: The characteristic function.}

\item{m}{integer: The number of samples.}

\item{m1}{Number of sampled permutations use to find a suitable transformation}
}
\value{
Estimated Shapley value of player \code{i}.
}
\description{
Approximates the Shapley value by ergodic sampling with the GreedyK2 algorithm.
}
\details{
Approximates the Shapley value by ergodic sampling. The GreedyK2 algorithm proposed by
Illes and Kerenyi (2022) is used to find the transformation, that generates samples
with negative covariance.
}
\examples{
ergodicSamplingGreedyK2(1, 10, gloveGameForSampling(1:5, 6:10), 20000, 10)
}
\references{
Illes, F. and Kerenyi, P. (2022) "Estimation of the Shapley value by ergodic sampling"
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
