% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromICoalitionToCoalition.R
\name{fromICoalitionToCoalition}
\alias{fromICoalitionToCoalition}
\title{from ICoalition To Coalition}
\usage{
fromICoalitionToCoalition(cardinality_n, iCoalition)
}
\arguments{
\item{cardinality_n}{Indicates the cardinality of the set of players \code{N}.}

\item{iCoalition}{Indicates the index of the coalition. This value should be greater than or equal to 1 and less than or equal to \code{2^n}.}
}
\value{
Numeric binary vector, of dimension \code{n}, which each component \code{i} indicates the belonging of player \code{i} to the coalition.
}
\description{
Access possible coalitions of players through index.
}
\details{
Based on: "Statistics and game theory: Estimating coalitional values in R software" (Alejandro Saavedra-Nieves, 2020)
}
\author{
Matthias Schöner
}
