% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromNumberToBinaryIndices.R
\name{fromNumberToBinaryIndices}
\alias{fromNumberToBinaryIndices}
\title{calculates the indices of the binary representation of the input number}
\usage{
fromNumberToBinaryIndices(num)
}
\arguments{
\item{num}{number}
}
\value{
vector of indices
}
\description{
supportfunction to calculate the indices of binary representations. Used for the calculation of permutations.
}
\examples{
# example code
# TODO: create example
}
