% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloveGameForSampling.R
\name{gloveGameForSampling}
\alias{gloveGameForSampling}
\title{Glove Game}
\usage{
gloveGameForSampling(L, R)
}
\arguments{
\item{L}{Set of players with a left glove.}

\item{R}{Set of players with a right glove,}
}
\value{
The characteristic function configured with the provided left and right glove lists that returns the number of matching glove pairs according to the provided set of players (S).
}
\description{
Returns the characteristic function of a glove game with the specified sets of left and right gloves.
}
\examples{
v <- gloveGameForSampling(L = 1:50, R = 51:100)
v(c(10, 90, 99, 70, 20))
}
\author{
Julian Musch

David Ulrich

Matthias Schöner

Tim Pollmann

Elias Wucher

Alexander Riessbeck

Robin Liebermann

Mithunan Mehanathan
}
