% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.R
\name{pre}
\alias{pre}
\title{Pre}
\usage{
pre(O, i)
}
\arguments{
\item{O}{Order}

\item{i}{integer: The player \code{i}.}
}
\value{
Vector of all predecessors of player \code{i} in order \code{O}.
}
\description{
Get predecessors of \code{i} in \code{O}.
Returns a vector of all predecessors of player \code{i} in order \code{O}.
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
