% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRandomSamplingWithReplacementAntithetic.R
\name{simpleRandomSamplingWithReplacementAntithetic}
\alias{simpleRandomSamplingWithReplacementAntithetic}
\title{A antithetic sampling procedure to estimate the Banzhaf value.}
\usage{
simpleRandomSamplingWithReplacementAntithetic(i, n, m, v)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}
}
\value{
Banzhaf value of player \code{i}.
}
\description{
simpleRandomSamplingWithReplacementAntithetic is a sampling methodology to estimate the
Banzhaf value for a specified player \code{i} and a specified TU game by antithetic sampling.
}
\details{
This algorithm approximates the Banzhaf value for a given player \code{i} by random
antithetic sampling. It is based on "Statistics and game theory: Estimating coalitional
values in R software" by A. Saavedra-Nieves (2020) and was extended to make use of antithetic
sampling by Staudacher and Pollmann (2023). The base algorithm can also be found in this package
under the name \code{simpleRandomSamplingWithReplacement}.
}
\examples{
simpleRandomSamplingWithReplacementAntithetic(1, 3, 200, gloveGameForSampling(1:2, 3:3))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Approximating cooperative games using antithetic sampling",  AppliedMath PLACEHOLDER VERSION PLACEHOLDER PAGES
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
