% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproOwenAndBanzhafOwen.R
\name{stApproOwenAndBanzhafOwen}
\alias{stApproOwenAndBanzhafOwen}
\title{A stratified sampling procedure to estimate coalitional values}
\usage{
stApproOwenAndBanzhafOwen(i, l, P, v, proportional = TRUE)
}
\arguments{
\item{i}{indicates player}

\item{l}{set total sample size}

\item{P}{A priori unions}

\item{v}{TU-game with a system of a priori unions}

\item{proportional}{boolean with true = propotional allocation and false = simple allocation procedure of samples}
}
\value{
approximation of the Banzaf-Owen value and Owen value based on stratified sampling
}
\description{
stApproOwenAndBanzhafOwen is a sampling methodology to estimate coalitional values, in particular the
Owen value and the Banzhaf–Owen value for a specified TU game with a system of a priori unions, based on stratified sampling.
Offering two allocation procedures for the samples.
Based on: "On stratified sampling for estimating coalitional values" (A. Saavedra-Nieves, 2022)
}
\examples{
print(stApproOwenAndBanzhafOwen(1, 10000, list(c(1, 2), c(3)), gloveGame(1:2, 3:3)))
print(stApproOwenAndBanzhafOwen(1, 10000, list(c(1, 2), c(3)), gloveGame(1:2, 3:3), FALSE)$Owen)

}
