% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproOwenAntithetic.R
\name{stApproOwenAntithetic}
\alias{stApproOwenAntithetic}
\title{A stratified antithetic sampling procedure to estimate coalitional values.}
\usage{
stApproOwenAntithetic(i, m, v, P)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}

\item{P}{list: A priori unions.}
}
\value{
Approximation of the Owen value based on stratified antithetic sampling.
}
\description{
stApproOwenAntithetic is a sampling methodology to estimate coalitional values, in particular the
Owen value for a specified TU game with a system of a priori unions, based on stratified and antithetic sampling.
}
\details{
The algorithm combines stratified with antithetic sampling. The base algorithm used
for the extension to antithetic sampling is based on the algorithm proposed by
A. Saavedra-Nieves (2022). This algorithm can be found under the name \code{stApproOwenAndBanzhafOwen}
in this package.
}
\examples{
stApproOwenAntithetic(1, 1000, gloveGameForSampling(1:2, 3:3), list(c(1, 2), c(3))))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Approximating cooperative games using antithetic sampling",  AppliedMath PLACEHOLDER VERSION PLACEHOLDER PAGES
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
