% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproShapley.R
\name{stApproShapley}
\alias{stApproShapley}
\title{Monte Carlo Simulation for calculating the Shapley value based on the paper "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation" by Castro et al. from 2017}
\usage{
stApproShapley(i, n, v, m)
}
\arguments{
\item{i}{Player}

\item{n}{Number of players}

\item{v}{Characteristic function}

\item{m}{Sample size}
}
\value{
Estimated Shapley value of player i
}
\description{
Monte Carlo Simulation for calculating the Shapley value based on the paper "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation" by Castro et al. from 2017
}
\examples{
stApproShapley(1, 10, gloveGame(1:5, 6:10), 1000)
}
