% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproShapleyAntithetic.R
\name{stApproShapleyAntithetic}
\alias{stApproShapleyAntithetic}
\title{A stratified antithetic sampling procedure to estimate the Shapley value}
\usage{
stApproShapleyAntithetic(i, n, m, v)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}
}
\value{
Estimated Shapley value of player \code{i}.
}
\description{
stApproShapleyAntithetic is a sampling methodology to estimate the
Shapley value for all players for a given TU game by stratified antithetic sampling.
}
\details{
This algorithm approximates the Shapley value for all players by stratified
antithetic sampling. It is based on "Improving polynomial estimation of the Shapley value by
stratified random sampling with optimum allocation" by Castro et al. (2017)
and was extended to make use of antithetic sampling by Staudacher and Pollmann (2023).
The base algorithm can also be found in this package under the name \code{stApproShapley}.
}
\examples{
stApproShapleyAntithetic(1, 3, 200, gloveGameForSampling(1:2, 3:3))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Approximating cooperative games using antithetic sampling",  AppliedMath
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
