% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetFromIndex.R
\name{subsetFromIndex}
\alias{subsetFromIndex}
\title{Gets the subset of players N (1:n) described by an index.}
\usage{
subsetFromIndex(n, index)
}
\arguments{
\item{n}{Indicates the cardinality of the set of players N}

\item{index}{A one based index describing the subset}
}
\value{
A subset from N described by index.
}
\description{
Gets the subset of players N (1:n) described by an index.
}
\examples{
S <- subsetFromIndex(5, 1) # empty set
S <- subsetFromIndex(5, 32) # all players
}
