% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematicSampling.R
\name{systematicSampling}
\alias{systematicSampling}
\title{Systematic Sampling}
\usage{
systematicSampling(i, n, m, v)
}
\arguments{
\item{i}{The player i}

\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{m}{The number of samples}

\item{v}{The characteristic function.}
}
\value{
The Banzhaf value of the specified player
}
\description{
Systematic sampling based on paper: "Statistics and game theory:
Estimating coalitional values in R software" (A. Saavedra-Nieves, 2020)
Algorithm 3
}
\examples{
print(systematicSampling(1, 10, 200, gloveGame(1:5, 6:10)))
}
\references{
Saavedra-Nieves, A. et al. (2020) "Statistics and game theory: Estimating coalitional values in R software ", Operations Research Letters 49 pp. 132
}
\author{
Alexander Rießbeck
}
