% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take.R
\name{take}
\alias{take}
\title{Take}
\usage{
take(source, index)
}
\arguments{
\item{source}{The source list}

\item{index}{Up-to where to include the elements}
}
\value{
A sub-list
}
\description{
Returns a sub-list of the provided source with the elements from the start up-to the specified index
}
\examples{
source <- "alfa,bravo,charlie,delta"
take(source, 2)
}
