% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageApproBanzhafOwen.R
\name{twoStageApproBanzhafOwen}
\alias{twoStageApproBanzhafOwen}
\title{A two-stage procedure to estimate the Banzhaf–Owen value}
\usage{
twoStageApproBanzhafOwen(i, lr, ls, v, P)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{lr}{number of samples of coalitions of unions; max of 2^(unions-1)}

\item{ls}{number of samples of coalitions in Pi for each element of sample;
max of 2^(pi-1)}

\item{v}{function: The characteristic function.}

\item{P}{list: A priori unions.}
}
\value{
approximation of the Banzaf-Owen value based on two-stage sampling
}
\description{
twoStageApproBanzhafOwen is a sampling method to approximate the Banzhaf-Owen value for
a specified TU game with a system of a priori unions, based on two-stage sampling.
}
\details{
First, the algorithm takes a sample of coalitions of unions
other than the one to which player i belongs to, without replacement. Then, for each element, the algorithm takes a sample of
coalitions without replacement of the union to which player i belongs.
Based on: "Sampling methods to estimate the Banzhaf–Owen value" (A. Saavedra-Nieves & M. G. Fiestras-Janeiro, 2020)
}
\examples{
print(twoStageApproBanzhafOwen(1, 2, 2, gloveGameForSampling(1:2, 3:3), list(c(1, 2), (3))))
}
\references{
Saavedra-Nieves, A. et al. (2020) "Statistics and game theory: Estimating coalitional values in R software ", Operations Research Letters 49 pp. 208-221
}
\author{
Matthias Schöner
}
