% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageStApproShapleyOptCorAntithetic.R
\name{twoStageStApproShapleyOptCorAntithetic}
\alias{twoStageStApproShapleyOptCorAntithetic}
\title{Two Stage Stratified ApproShapley with Optimum Allocation (Corrected) with Antithetic Sampling}
\usage{
twoStageStApproShapleyOptCorAntithetic(n, v, m)
}
\arguments{
\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{v}{function: The characteristic function.}

\item{m}{integer: The number of samples.}
}
\value{
A list of Shapley values of every player.
}
\description{
Calculates the Shapley value for each player by using stratification and antithetic sampling. The
sample size per stratum is based on each stratums estimated variance.
}
\details{
\code{twoStageStApproShapleyOptCorAntithetic} is a sampling procedure to estimate the Shapley value for cooperative games.
In the first stage, the optimal allocation is calculated and executed in the second stage. Using
the original algorithm \code{twoStageStApproShapleyOpt} from Castro et al. (2017), it may occur that
more samples than specified are used. If this is the case, this algorithm here adjusts the sample sizes proportionally.
Furthermore, in comparison to the original \code{twoStageStApproShapleyOpt}, this algorithm
was extended to make use of antithetic sampling by Staudacher and Pollmann (2023).
}
\examples{
# sample a airport game with 100 players
\donttest{
costs <- buildAirportCostVector(list(
  c(1, 8), c(2, 12), c(3, 6), c(4, 14), c(5, 8), c(6, 9),
  c(7, 13), c(8, 10), c(9, 10), c(10, 10)
))
v <- airportGameForSampling(costs)
Sh <- twoStageStApproShapleyOptCorAntithetic(length(costs), v, 100000)
print(Sh)
}
print(twoStageStApproShapleyOptCorAntithetic(3, gloveGameForSampling(1:2, 3:3), 1000))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Assessing antithetic sampling for the Shapley, Banzhaf and Owen values", Preprint. Submitted to AppliedMath. Under Review.
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
