% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageStApproShapleyOptEstimateSamples.R
\name{twoStageStApproShapleyOptEstimateSamples}
\alias{twoStageStApproShapleyOptEstimateSamples}
\title{Two Stage St Appro Shapley Opt Estimate Samples}
\usage{
twoStageStApproShapleyOptEstimateSamples(n, v, error, probability)
}
\arguments{
\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{v}{function: The characteristic function.}

\item{error}{The error (difference between actual shapley value and estimated
shapley value) we do not want to exceed.}

\item{probability}{How likely it is that we stay below 'error'.}
}
\value{
An estimate for m/the sample size (upper bound).
}
\description{
Estimates the sample count m that should be used with
'twoStageStApproShapleyOpt', based on the given error and probability.
The returned m leads to an estimation of shapley values where the difference
between estimated Shapley value and actual Shapley value is smaller than or
equal to 'error' with a probability of 'probability'.
}
\details{
Proposition 4.2
}
\examples{
v <- gloveGameForSampling(L = 1:5, R = 6:10)
m <- twoStageStApproShapleyOptEstimateSamples(10, v, 0.2, 0.8)
}
\references{
CASTRO, J. et al. (2017) "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation", Computers and Operations Research 82, pp. 184
}
\author{
David Ulrich
}
