% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approOwenAntithetic.R
\name{approOwenAntithetic}
\alias{approOwenAntithetic}
\title{An antithetic sampling procedure to estimate the Owen value.}
\usage{
approOwenAntithetic(i, m, v, P)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}

\item{P}{list: A priori unions.}
}
\value{
Owen value of player \code{i}.
}
\description{
approOwenAntithetic is a sampling methodology to estimate the
Owen value for a given player \code{i} and a specified TU game with a system of
a priori unions by antithetic sampling.
}
\details{
This algorithm approximates the Owen value for one player by random antithetic
sampling. It is based on "Estimation of the Owen Value Based on Sampling" by Alejandro
Saavedra-Nieves et al. (2018) and was extended to make use of antithetic sampling by Staudacher and
Pollmann (2023).
}
\examples{
approOwenAntithetic(1, 1000, gloveGameForSampling(1:2, 3:3), list(c(1, 2), c(3)))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Approximating cooperative games using antithetic sampling",  AppliedMath PLACEHOLDER VERSION PLACEHOLDER PAGES
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
