% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approShapley.R
\name{approShapley}
\alias{approShapley}
\title{Appro Shapley}
\usage{
approShapley(n, m, v)
}
\arguments{
\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}
}
\value{
A list of Shapley values of every player.
}
\description{
Approximates the Shapley value by using a simple Monte Carlo simulation.
}
\details{
Appro Shapley is a sampling method, which samples orders of the players.
In each of these orders the marginal contribution of each player in the
current order will be calculated to estimate the real Shapley value.
Note that it is possible that the provided sample size is not divisible by
the number of players. In that case the remaining samples \code{m \%\% n} will
not be used. This preserves the efficiency (i.e. the sum of the result vector
is 1) of the algorithm.
Based on: "Polynomial calculation of the Shapley value based on sampling"
(Javier Castro Et al., 2008)
}
\examples{
approShapley(10, 10000, gloveGameForSampling(1:5, 6:10))
\donttest{
# sample an airport game as described in section 4.3 of the underlying paper
costs <- buildAirportCostVector(list(
  c(1, 8), c(2, 12), c(3, 6), c(4, 14), c(5, 8), c(6, 9),
  c(7, 13), c(8, 10), c(9, 10), c(10, 10)
))
v <- airportGameForSampling(costs)
Sh <- approShapley(length(costs), 1000000, v)
}
}
\references{
CASTRO, J. et al. (2008) "Polynomial calculation of the Shapley value based on sampling", Computers and Operations Research 36, pp. 1727
}
\author{
Julian Musch

David Ulrich

Tim Pollmann

Elias Wucher

Matthias Schöner
}
