% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_conf}
\alias{check_conf}
\title{Check parameter \code{conf}}
\usage{
check_conf(conf)
}
\arguments{
\item{conf}{numeric: Confidence (probability) of the true (exact) value being in the provided interval.}
}
\value{
None
}
\description{
Check parameter \code{conf} for validity.
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
