% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_m}
\alias{check_m}
\title{Check parameter \code{m}}
\usage{
check_m(m, max_value = NULL, bigz_allowed = FALSE)
}
\arguments{
\item{m}{integer: The number of samples.}

\item{max_value}{Maximal value \code{m} can take.}

\item{bigz_allowed}{Determine if big ints from \code{gmp} are allowed.}
}
\value{
None
}
\description{
Check parameter \code{m} for validity.
}
\author{
Alexander Riessbeck
}
