% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_m}
\alias{check_m}
\title{Check parameter m}
\usage{
check_m(m, max_value = NULL)
}
\arguments{
\item{m}{The number of samples}

\item{max_value}{Maximal value m can take}
}
\value{
Returns NULL
}
\description{
Check parameter m for validity.
}
\author{
Alexander Rießbeck
}
