% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_m_n}
\alias{check_m_n}
\title{Check parameters m and n}
\usage{
check_m_n(m, n, m_max = NULL)
}
\arguments{
\item{m}{The number of samples}

\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{m_max}{Maximal value m can take}
}
\value{
Returns NULL
}
\description{
Check parameters m and n for validity.
}
\author{
Tim Pollmann
}
