% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_m_n}
\alias{check_m_n}
\title{Check parameters \code{m} and \code{n}}
\usage{
check_m_n(m, n, m_max = NULL, bigz_allowed = FALSE)
}
\arguments{
\item{m}{integer: The number of samples.}

\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{m_max}{Maximal value \code{m} can take.}

\item{bigz_allowed}{Determines if bigz are allowed.}
}
\value{
None
}
\description{
Check parameters \code{m} and \code{n} for validity.
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
