% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_n_i}
\alias{check_n_i}
\title{Check parameter n and i}
\usage{
check_n_i(n, i)
}
\arguments{
\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{i}{The player i}
}
\value{
Returns
}
\description{
Check parameter n and i for validity with the intended functions for that purpose.
Additionally it checks if i is smaller or equal n, because player i must be in the quantity
generated by n.
}
\author{
Elias Wucher

Alexander Rießbeck
}
