% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_natural_number}
\alias{check_natural_number}
\title{Check if the parameter is an natural number.}
\usage{
check_natural_number(parameter)
}
\arguments{
\item{parameter}{Parameter to check}
}
\value{
None
}
\description{
Checks if the value of the parameter is an natural number. Note: The
datatype will not be checked. It will only be checked if the value of the
parameter is greater than 0 and not a point number.
}
\author{
Tim Pollmann
}
