% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_P}
\alias{check_P}
\title{Check parameter \code{P}}
\usage{
check_P(P)
}
\arguments{
\item{P}{list: A priori unions.}
}
\value{
None
}
\description{
Check parameter \code{P} for validity. Parameter \code{P} must not be \code{NULL}, must be a
list of vectors or a list of lists for the priori unions. The priori unions
must be numeric and positive. Additionally, the priori unions have to be disjunct.
}
\author{
Elias Wucher
}
