% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_P_i}
\alias{check_P_i}
\title{Check parameter \code{P} and \code{i}}
\usage{
check_P_i(P, i)
}
\arguments{
\item{P}{list: A priori unions.}

\item{i}{integer: The player \code{i}.}
}
\value{
None
}
\description{
Check parameter \code{P} and \code{i} for validity with the intended functions for that purpose.
Additionally, check if player \code{i} is element of a priori union in \code{P}.
}
\author{
Elias Wucher
}
