% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_positive_number}
\alias{check_positive_number}
\title{Check if the parameter is a number and positive}
\usage{
check_positive_number(parameter)
}
\arguments{
\item{i}{The player i}
}
\value{
Returns
}
\description{
Check if the parameter is a number and positive. It can be used for the parameter
n, i, m and w.
}
\author{
Elias Wucher

Alexander Rießbeck
}
