% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_v}
\alias{check_v}
\title{Check parameter \code{v}}
\usage{
check_v(v)
}
\arguments{
\item{v}{function: The characteristic function.}
}
\value{
None
}
\description{
Check parameter \code{v} for validity. \code{v} must not be \code{NULL} and it has to be a function.
}
\author{
Elias Wucher
}
