% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{check_v}
\alias{check_v}
\title{Check parameter v}
\usage{
check_v(v)
}
\arguments{
\item{v}{function: the characteristic function.}
}
\value{
Returns

example
check_v(function(x)x)
}
\description{
Check parameter v for validity. V must not be NULL and it has to be a function
}
\author{
Elias Wucher
}
