% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCommonParameter.R
\name{checkCommonParameter}
\alias{checkCommonParameter}
\title{Checks the common parameters for their validity}
\usage{
checkCommonParameter(v, n = NULL, m = NULL, i = NULL, P = NULL, w = NULL)
}
\arguments{
\item{v}{The characteristic function.}

\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{m}{The number of samples}

\item{i}{The player i}

\item{P}{A priori unions}

\item{w}{Width of the confidence interval}
}
\value{
Returns true if the parameters are valid and false if they aren't
}
\description{
Checks the common parameters for their validity.

The parameters must have the following properties:
 - The number of players can't be negative oder zero
 -
}
\author{
Elias Wucher
}
