% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitionFromIndex.R
\name{coalitionFromIndex}
\alias{coalitionFromIndex}
\title{Get a coalition from an index}
\usage{
coalitionFromIndex(superset, index)
}
\arguments{
\item{superset}{The superset from which to get a specific coalition described
by the parameter index.}

\item{index}{Index describing the coalition/subset. Valid values are in the
range from 1 to including \code{2^len(superset)}.}
}
\value{
A coalition/subset from superset described by index
}
\description{
Gets the coalition/subset of superset described by index.
}
\author{
Alexander Riessbeck

Matthias Schoener \href{https://orcid.org/0009-0003-4007-4362}{0009-0003-4007-4362}

David Ulrich
}
