% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBanzhaf.R
\name{confidenceBanzhaf}
\alias{confidenceBanzhaf}
\title{Monte Carlo Simulation for calculating the Banzhaf value in a simple game based on the paper "Approximating power indices: theoretical and empirical analysis" by Bachrach et al. from 2010}
\usage{
confidenceBanzhaf(i, n, v, conf, w)
}
\arguments{
\item{i}{Player}

\item{n}{Number of players}

\item{v}{Characteristic function}

\item{conf}{Confidence of the true Banzhaf value being in the confidence interval}

\item{w}{Width of the confidence interval}
}
\value{
Confidence interval as a vector of size 2
}
\description{
Monte Carlo Simulation for calculating the Banzhaf value in a simple game based on the paper "Approximating power indices: theoretical and empirical analysis" by Bachrach et al. from 2010
}
\examples{
confidenceBanzhaf(3, 4, weightedVotingGame(c(1, 1, 2, 3), 1 / 2), 0.95, 0.01)
}
