% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceShapleyShubik.R
\name{confidenceShapleyShubik}
\alias{confidenceShapleyShubik}
\title{Confidence Shapley Shubik}
\usage{
confidenceShapleyShubik(i, n, v, conf, w)
}
\arguments{
\item{i}{The player i}

\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{v}{The characteristic function.}

\item{conf}{Confidence (probability) of the true (exact) value being in the interval}

\item{w}{Width of the confidence interval}
}
\value{
Confidence interval as a vector of size 2
}
\description{
Monte Carlo Simulation for calculating the Shapley-Shubik power index in a simple game.
}
\details{
This function calculates an approximation of the Shapley-Shubik value.
You do not have to provide a sample size. The needed sample size is calculated
inside the function based on the given confidence and interval width.
The interval width is the width of the interval of the approximated Shapley-Shubik value
(accuracy) while the confidence is the probability of the true (exact) Shapley-Shubik
value being in this interval. Decreasing the interval width and/ or increasing
the confidence will lead to an increase of the required samples and a longer
computation time.
The method is based on the paper "Approximating power indices: theoretical an
empirical analysis" by Bachrach et al. from 2010.
}
\examples{
confidenceShapleyShubik(3, 4, weightedVotingGameForSampling(c(1, 1, 2, 3), 1 / 2), 0.95, 0.01)
}
\references{
Bachrach, Y. et al. (2010) "Approximating power indices: theoretical and empirical analysis", Auton Agent Multi-Agent Syst 20, pp. 105-122
}
\author{
Tim Pollmann
}
