% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.R
\name{pre}
\alias{pre}
\title{Pre}
\usage{
pre(O, i)
}
\arguments{
\item{O}{Order}

\item{i}{Player}
}
\value{
Vector of all predecessors of player i in order O
}
\description{
Get predecessors of i in O
Returns a vector of all predecessors of player i in order O
}
\examples{
pre(c(1, 4, 2, 3), 2)
}
