% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRandomSamplingWithoutReplacement.R
\name{simpleRandomSamplingWithoutReplacement}
\alias{simpleRandomSamplingWithoutReplacement}
\title{Simple random sampling without replacement}
\usage{
simpleRandomSamplingWithoutReplacement(i, n, m, v)
}
\arguments{
\item{i}{integer: the player i}

\item{n}{integer: indicates the cardinality of the set of players N (The number of players).}

\item{m}{integer or bigz: the number of samples}

\item{v}{function: the characteristic function.}
}
\value{
The Banzhaf value of the specified player
}
\description{
Simple random sampling without replacement is a method to
approximate the Banzhaf value for a specified TU game.
}
\details{
First, the algorithm takes a sample without replacement of all
players without the player i. Then the algorithm checks the value the player
adds to the sample. These two steps are repeated for the number of samples.

Based on the paper: "Statistics and game theory: Estimating coalitional
values in R software" (A. Saavedra-Nieves, 2020) Algorithm 2
}
\examples{
print(simpleRandomSamplingWithoutReplacement(1, 10, 200, gloveGameForSampling(1:5, 6:10)))
# --------------
\dontrun{
# very long runtime
m <- as.bigz(.Machine$integer.max) + 1
print(simpleRandomSamplingWithoutReplacement(1, 100, m, gloveGameForSampling(1:50, 51:100)))
}
}
\references{
Saavedra-Nieves, A. et al. (2020) "Statistics and game theory: Estimating coalitional values in R software ", Operations Research Letters 49 pp. 131
}
\author{
Alexander Rießbeck

Robin Liebermann

Mithunan Mehanathan
}
