% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproOwenAndBanzhafOwen.R
\name{stApproOwenAndBanzhafOwen}
\alias{stApproOwenAndBanzhafOwen}
\title{A stratified sampling procedure to estimate coalitional values.}
\usage{
stApproOwenAndBanzhafOwen(i, m, v, P, proportional = TRUE)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{m}{integer or bigz: the number of samples.}

\item{v}{function: The characteristic function.}

\item{P}{list: A priori unions.}

\item{proportional}{boolean with true = propotional allocation and false = simple allocation procedure of samples}
}
\value{
Approximation of the Banzaf-Owen value and Owen value based on stratified sampling.
}
\description{
stApproOwenAndBanzhafOwen is a sampling methodology to estimate coalitional values, in particular the
Owen value and the Banzhaf–Owen value for a specified TU game with a system of a priori unions, based on stratified sampling.
Offering two allocation procedures for the samples.
}
\details{
The algorithm loops through a strata of \code{k} unions and \code{h} players.
Firstly, the algorithm takes with replacement a sample of coalitions of size \code{k} composed by unions other than the
one to which player \code{i} belongs. Secondly, for each element of the first sample, the algorithm takes with replacement a coalition
of players in the union of player \code{i} other than player \code{i} itself. This sample is used to calculate the Owen and Banzhaf-Owen values.
If \code{proportional} is set, the samples are allocated proportionally to the weight of the strata. Otherwise, they are evenly distributed.
Based on: "On stratified sampling for estimating coalitional values" (A. Saavedra-Nieves, 2022)

This algorithm supports BigQ from GMP.
}
\examples{
print(stApproOwenAndBanzhafOwen(1, 1000, gloveGameForSampling(1:2, 3:3), list(c(1, 2), c(3))))
print(stApproOwenAndBanzhafOwen(
  1, 1000, gloveGameForSampling(1:2, 3:3),
  list(c(1, 2), c(3)), FALSE
)$Owen)

}
\references{
Saavedra-Nieves, A. et al. (2022) "On stratified sampling for estimating coalitional values", Annals of Operations Research 320 pp. 5
}
\author{
Matthias Schoener \href{https://orcid.org/0009-0003-4007-4362}{0009-0003-4007-4362}
}
