% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproOwenAndBanzhafOwen.R
\name{stApproOwenAndBanzhafOwen}
\alias{stApproOwenAndBanzhafOwen}
\title{A stratified sampling procedure to estimate coalitional values}
\usage{
stApproOwenAndBanzhafOwen(i, m, v, P, proportional = TRUE)
}
\arguments{
\item{i}{The player i}

\item{m}{The number of samples}

\item{v}{The characteristic function.}

\item{P}{A priori unions}

\item{proportional}{boolean with true = propotional allocation and false = simple allocation procedure of samples}
}
\value{
approximation of the Banzaf-Owen value and Owen value based on stratified sampling
}
\description{
stApproOwenAndBanzhafOwen is a sampling methodology to estimate coalitional values, in particular the
Owen value and the Banzhaf–Owen value for a specified TU game with a system of a priori unions, based on stratified sampling.
Offering two allocation procedures for the samples.
}
\details{
The algorithm loops through a strata of k unions and h players.
Firstly, the algorithm takes with replacement a sample of coalitions of size k composed by unions other than the
one to which player i belongs. Secondly, for each element of the first sample, the algorithm takes with replacement a coalition
of players in the union of player i other than player i itself. This sample is used to calculate the Owen and Banzhaf-Owen values.
If propotional is set, the samples are allocated proportionally to the weight of the strata. Otherwise they are evenly distributed.
Using BigInt(gmp) for m is possible.
Based on: "On stratified sampling for estimating coalitional values" (A. Saavedra-Nieves, 2022)
}
\examples{
print(stApproOwenAndBanzhafOwen(1, 1000, gloveGameForSampling(1:2, 3:3), list(c(1, 2), c(3))))
print(stApproOwenAndBanzhafOwen(1, 1000, gloveGameForSampling(1:2, 3:3), list(c(1, 2), c(3)), FALSE)$Owen)

}
\references{
Saavedra-Nieves, A. et al. (2022) "On stratified sampling for estimating coalitional values", Annals of Operations Research 320 pp. 5
}
\author{
Matthias Schöner
}
