% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproOwenAntithetic.R
\name{stApproOwenAntithetic}
\alias{stApproOwenAntithetic}
\title{Owen value approximation by stratified antithetic sampling}
\usage{
stApproOwenAntithetic(i, m, v, P)
}
\arguments{
\item{i}{integer: The player \code{i}.}

\item{m}{integer: The number of samples.}

\item{v}{function: The characteristic function.}

\item{P}{list: A priori unions.}
}
\value{
Owen value of player \code{i}.
}
\description{
\code{stApproOwenAntithetic} is a sampling algorithm to estimate the
Owen value for a given player \code{i} for a specified TU game with a system of
a priori unions by using stratified antithetic sampling.
}
\details{
This algorithm approximates the Owen value for one player by stratified
antithetic sampling. It is based on "On stratified sampling for estimating coalitional
values" by A. Saavedra-Nieves (2022) and was extended to make use of antithetic
sampling by Staudacher and Pollmann (2023). The base algorithm can also be found in this package
under the name \code{stApproOwenAndBanzhafOwen}.
Note that the sample allocation over the strata is proportional to the weights of the strata.
We refer to Staudacher and Pollmann (2023) for more details, especially regarding an edge
case that occurs when the number precoalitions and the number of players in player
i's union are both odd.
}
\examples{
stApproOwenAntithetic(1, 1000, gloveGameForSampling(1:2, 3:3), list(c(1, 2), c(3)))
}
\references{
Staudacher, J. and Pollmann, T. (2023) "Approximating cooperative games using antithetic sampling",  AppliedMath
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
