% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproShapley.R
\name{stApproShapley}
\alias{stApproShapley}
\title{Stratified ApproShapley}
\usage{
stApproShapley(n, v, m)
}
\arguments{
\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{v}{function: The characteristic function.}

\item{m}{integer: The number of samples.}
}
\value{
A list of Shapley values of every player.
}
\description{
Monte Carlo Simulation for calculating the Shapley value with stratification.
}
\details{
This function uses stratification to approximate the Shapley value. A stratum
is created for each player at each possible position. The estimated Shapley value
is the mean over the estimated Shapley values of all strata.
}
\examples{
stApproShapley(10, gloveGameForSampling(1:5, 6:10), 1000)
}
\references{
CASTRO, J. et al. (2017) "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation", Computers and Operations Research 82, pp. 183
}
\author{
Tim Pollmann \href{https://orcid.org/0009-0008-4627-7925}{0009-0008-4627-7925}
}
