% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproShapley.R
\name{stApproShapley}
\alias{stApproShapley}
\title{Stratified Appro Shapley}
\usage{
stApproShapley(i, n, v, m)
}
\arguments{
\item{i}{The player i}

\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{v}{The characteristic function.}

\item{m}{The number of samples}
}
\value{
Estimated Shapley value of player i
}
\description{
Monte Carlo Simulation for calculating the Shapley value based on the paper "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation" by Castro et al. from 2017
}
\examples{
stApproShapley(1, 10, gloveGame(1:5, 6:10), 1000)
}
\references{
CASTRO, J. et al. (2017) "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation", Computers and Operations Research 82, pp. 5
}
\author{
Tim Pollman
}
