% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematicSampling.R
\name{systematicSampling}
\alias{systematicSampling}
\title{Systematic sampling based on paper: "Statistics and game theory: Estimating coalitional values in R software" (A. Saavedra-Nieves, 2020) Algorithm 3}
\usage{
systematicSampling(all_players_N, player_i, sampling_size_l, game_v)
}
\arguments{
\item{all_players_N}{List of all players}

\item{player_i}{Index of the player}

\item{sampling_size_l}{Amount of samples with replacement}

\item{game_v}{Function of the TU-game}
}
\value{
Banzhaf value
}
\description{
Systematic sampling based on paper: "Statistics and game theory: Estimating coalitional values in R software" (A. Saavedra-Nieves, 2020) Algorithm 3
}
\examples{
print(systematicSampling(1:10, 1, 200, gloveGame(1:5, 6:10)))
}
