% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematicSampling.R
\name{systematicSampling}
\alias{systematicSampling}
\title{Systematic Sampling}
\usage{
systematicSampling(i, n, m, v)
}
\arguments{
\item{i}{integer: the player i}

\item{n}{integer: indicates the cardinality of the set of players N (The number of players).}

\item{m}{integer or bigz: the number of samples}

\item{v}{function: the characteristic function.}
}
\value{
The Banzhaf value of the specified player
}
\description{
Systematic sampling is a method to estimate the Banzhaf-Owen
  value for a specified TU game.
}
\details{
First, the algorithm systematically creates samples of players
without the player i by incrementally taking samples out of a sequence of all
possible combinations of players. Then the algorithm checks the value the
player adds to all the samples.

Based on the paper: "Statistics and game theory: Estimating coalitional
values in R software" (A. Saavedra-Nieves, 2020) Algorithm 3
}
\examples{
print(systematicSampling(1, 10, 200, gloveGameForSampling(1:5, 6:10)))
# --------------
\dontrun{
# very long runtime
m <- as.bigz(.Machine$integer.max) + 1
print(simpleRandomSamplingWithoutReplacement(1, 100, m, gloveGameForSampling(1:50, 51:100)))
}
}
\references{
Saavedra-Nieves, A. et al. (2020) "Statistics and game theory: Estimating coalitional values in R software ", Operations Research Letters 49 pp. 132
}
\author{
Alexander Rießbeck

Robin Liebermann

Mithunan Mehanathan
}
