% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageApproShapleyEff.R
\name{twoStageApproShapleyEff}
\alias{twoStageApproShapleyEff}
\title{Described in "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation" on page 183
Calculates the shapley value based on the twoStageStApproShapleyOpt algorithm and the common approShapley
with the extension, that the algorithm calculates the value Efficient}
\usage{
twoStageApproShapleyEff(n, v, sample_size_eff, sample_size_opt)
}
\arguments{
\item{n}{The number of players}

\item{v}{The characteristic function}

\item{sample_size_eff}{The amount of samples for the common approShapley algorithm.}

\item{sample_size_opt}{The amount of samples for the twoStageStApproShapleyOpt algorithm}
}
\value{
The efficient shapley value of each player
}
\description{
Described in "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation" on page 183
Calculates the shapley value based on the twoStageStApproShapleyOpt algorithm and the common approShapley
with the extension, that the algorithm calculates the value Efficient
}
\examples{
# sample a airport game with 100 players
costs <- buildAirportCostVector(list(
  c(1, 8), c(2, 12), c(3, 6), c(4, 14), c(5, 8), c(6, 9),
  c(7, 13), c(8, 10), c(9, 10), c(10, 10)
))
v <- airportGame(costs)
Sh <- twoStageStApproShapleOpt(length(costs), v, 100000)
}
