% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageApproShapleyEff.R
\name{twoStageApproShapleyEff}
\alias{twoStageApproShapleyEff}
\title{Two Stage Appro Shapley Eff}
\usage{
twoStageApproShapleyEff(n, v, sample_size_eff, sample_size_opt)
}
\arguments{
\item{n}{The number of players.}

\item{v}{The characteristic function.}

\item{sample_size_eff}{The amount of samples for the common approShapley algorithm.}

\item{sample_size_opt}{The amount of samples for the twoStageStApproShapleyOpt algorithm}
}
\value{
The shapley value of each player with the addition, that sum(sh_i) = v(N) (efficient)
}
\description{
Described in "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation" on page 183
Calculates the shapley value based on the twoStageStApproShapleyOpt algorithm and the common approShapley
with the extension, that the algorithm calculates the value Efficient
}
\examples{
# sample a glove Game with 100 players (|L|=50 => |R|=50)
L <- 1:50
R <- 51:100
Sh <- twoStageApproShapleyEff(100, gloveGame(L, R), 10000, 100000)
}
\references{
CASTRO, J. et al. (2017) "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation", Computers and Operations Research 82, TODOe
}
