% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageStApproShapleyOpt.R
\name{twoStageStApproShapleyOpt}
\alias{twoStageStApproShapleyOpt}
\title{Calculates the shapley value for each player with stratification and
sample size per strata based on each stratums variance}
\usage{
twoStageStApproShapleyOpt(n, v, min_sample_size, first_stage_size = 0.5)
}
\arguments{
\item{n}{The number of players}

\item{v}{The characteristic function}

\item{min_sample_size}{The amount of samples that should be taken.
Based on the variances of each stratum it is likely to happen, that more
samples are used than specified in min_sample_size. The actual sample size
depends on the problem.}

\item{first_stage_size}{The amount of samples to use for the first stage as
a ratio. Usually 0.5 (half of all samples).}
}
\value{
The shapley value of each player
}
\description{
Calculates the shapley value for each player with stratification and
sample size per strata based on each stratums variance
}
\examples{
# sample a airport game with 100 players
costs <- buildAirportCostVector(list(
  c(1, 8), c(2, 12), c(3, 6), c(4, 14), c(5, 8), c(6, 9),
  c(7, 13), c(8, 10), c(9, 10), c(10, 10)
))
v <- airportGame(costs)
Sh <- twoStageStApproShapleyOpt(length(costs), v, 100000)
}
