% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageStApproShapleyOptCor.R
\name{twoStageStApproShapleyOptCor}
\alias{twoStageStApproShapleyOptCor}
\title{Two Stage St Appro Shapley Opt Cor}
\usage{
twoStageStApproShapleyOptCor(n, v, min_sample_size)
}
\arguments{
\item{n}{Indicates the cardinality of the set of players N (The number of players).}

\item{v}{The characteristic function.}

\item{min_sample_size}{The amount of samples that should be taken.
Based on the variances of each stratum it is likely to happen, that more
samples are used than specified in min_sample_size. These samples are proportionally
adjusted-}
}
\value{
A list of Shapley values of every player
}
\description{
Calculates the shapley value for each player with stratification and
sample size per strata based on each stratums variance samples size adjusted
}
\details{
twoStageStApproShapleyOptCor is a sampling procedure to estimate the Shapley value for cooperative games.
Based on: "Improving polynomial estimation of the Shapley value by stratified
random sampling with optimum allocation" (J. Castro Et al., 2017)
}
\examples{
# sample a airport game with 100 players
costs <- buildAirportCostVector(list(
  c(1, 8), c(2, 12), c(3, 6), c(4, 14), c(5, 8), c(6, 9),
  c(7, 13), c(8, 10), c(9, 10), c(10, 10)
))
v <- airportGameForSampling(costs)
Sh <- twoStageStApproShapleyOptCor(length(costs), v, 100000)
print(Sh)
}
\references{
CASTRO, J. et al. (2017) "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation", Computers and Operations Research 82, pp. 182
}
\author{
Matthias Schöner
}
