% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageStApproShapleyOptCor.R
\name{twoStageStApproShapleyOptCor}
\alias{twoStageStApproShapleyOptCor}
\title{Two Stage St Appro Shapley Opt Cor}
\usage{
twoStageStApproShapleyOptCor(n, v, m)
}
\arguments{
\item{n}{integer: Indicates the cardinality of the set of players \code{N} (The number of players).}

\item{v}{function: The characteristic function.}

\item{m}{integer: The number of samples.}
}
\value{
A list of Shapley values of every player.
}
\description{
Calculates the shapley value for each player with stratification and
sample size per strata based on each stratums variance samples size adjusted.
}
\details{
twoStageStApproShapleyOptCor is a sampling procedure to estimate the Shapley value for cooperative games.
In the first stage, the optimal allocation is calculated and executed in the second stage. It may occur that
more samples would be used than specified. If this is the case, the algorithm adjusts the samples proportionally
to the given sample size. The algorithm twoStageStApproShapleyOpt does not correct the sample size.
Modified according to: "Improving polynomial estimation of the Shapley value by stratified
random sampling with optimum allocation" (J. Castro Et al., 2017)
}
\examples{
# sample a airport game with 100 players
\donttest{
costs <- buildAirportCostVector(list(
  c(1, 8), c(2, 12), c(3, 6), c(4, 14), c(5, 8), c(6, 9),
  c(7, 13), c(8, 10), c(9, 10), c(10, 10)
))
v <- airportGameForSampling(costs)
Sh <- twoStageStApproShapleyOptCor(length(costs), v, 100000)
print(Sh)
}
print(twoStageStApproShapleyOptCor(3, gloveGameForSampling(1:2, 3:3), 1000))
}
\references{
CASTRO, J. et al. (2017) "Improving polynomial estimation of the Shapley value by stratified random sampling with optimum allocation", Computers and Operations Research 82, pp. 182
}
\author{
Matthias Schoener \href{https://orcid.org/0009-0003-4007-4362}{0009-0003-4007-4362}

Tim Pollmann
}
