% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedVotingGameForSampling.R
\name{weightedVotingGameForSampling}
\alias{weightedVotingGameForSampling}
\title{Weighted Voting Game}
\usage{
weightedVotingGameForSampling(weights, quota)
}
\arguments{
\item{weights}{The weights of the players}

\item{quota}{The required relative quota}
}
\value{
1 if sum of weighted votes >= quota; 0 otherwise
}
\description{
Returns the characteristic function of a weighted voting game with the specified weights and quota
}
\examples{
v <- weightedVotingGameForSampling(c(1, 2, 3, 4, 5, 4, 3, 2, 1, 10), 1 / 2)
v(c(1:10))
}
\author{
Julian Musch

Matthias Schöner

Tim Pollman

David Ulrich

Alexander Rießbeck

Robin Liebermann

Mithunan Mehanathan
}
